Move QueryResult to cache crate
